+++ examples/aot/fast_inverse/inverse_builder.py
@@ -9,11 +9,11 @@ MAX_MATRIX_SIZE = 128
 
 
 def meta_data():
-    in_dtype = pto.float16
-    # PTO-DSL tile.mov currently requires identical src/dst element dtype.
-    # The source C++ uses ACC(fp32) -> MAT(fp16) TMOV, which cannot be modeled
-    # directly here, so we keep the full tile pipeline in fp16.
-    out_dtype = pto.float16
+    # PTO-DSL currently enforces same-dtype tile.mov; this means we cannot
+    # model ACC(fp32) -> MAT(fp16) feedback used in hand-written C++.
+    # Keep all matrix tiles in fp32 so TMOV feedback paths verify and compile.
+    in_dtype = pto.float32
+    out_dtype = pto.float32
     i32 = pto.int32
 
     in_ptr_type = pto.PtrType(in_dtype)
